feat: 포모도로 API 1차 완료

+++ src/main/java/com/fivlo/fivlo_backend/domain/pomodoro/service/PomodoroService.java
@@ -1,17 +1,20 @@
 package com.fivlo.fivlo_backend.domain.pomodoro.service;
 
-import com.fivlo.fivlo_backend.domain.pomodoro.dto.PomodoroCreateRequest;
-import com.fivlo.fivlo_backend.domain.pomodoro.dto.PomodoroGoalResponse;
+import com.fivlo.fivlo_backend.domain.pomodoro.dto.*;
 import com.fivlo.fivlo_backend.domain.pomodoro.entity.PomodoroGoal;
+import com.fivlo.fivlo_backend.domain.pomodoro.entity.PomodoroSession;
 import com.fivlo.fivlo_backend.domain.pomodoro.repository.PomodoroGoalRepository;
 import com.fivlo.fivlo_backend.domain.pomodoro.repository.PomodoroSessionRepository;
+import com.fivlo.fivlo_backend.domain.pomodoro.dto.CoinByPomodoroSessionResponse;
 import com.fivlo.fivlo_backend.domain.user.entity.User;
 import com.fivlo.fivlo_backend.domain.user.repository.UserRepository;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.time.LocalDate;
 import java.util.List;
 import java.util.NoSuchElementException;
 
@@ -23,6 +26,7 @@ public class PomodoroService {
     private final PomodoroGoalRepository pomodoroGoalRepository;
     private final UserRepository userRepository;
 
+    /// //// 미완성 !*******************
     @Transactional(readOnly = true)
     public PomodoroGoalResponse findPomodoroGoals(Long id) {
         List<PomodoroGoal> pomodoroGoalList = pomodoroGoalRepository.findByUserId(id);
@@ -30,7 +34,7 @@ public class PomodoroService {
     }
 
     @Transactional
-    public Long create(Long id, @Valid PomodoroCreateRequest dto) {
+    public Long createGoal(Long id, @Valid PomodoroGoalCreateRequest dto) {
         User findUser = userRepository.findById(id)
                 .orElseThrow(() -> new NoSuchElementException("사용자를 찾을 수 없습니다."));
 
@@ -45,7 +49,7 @@ public class PomodoroService {
     }
 
     @Transactional
-    public String update(Long id, Long goalId, @Valid PomodoroCreateRequest dto) {
+    public String update(Long id, Long goalId, @Valid PomodoroGoalCreateRequest dto) throws AccessDeniedException {
 
         // 포모도로 목표 조회
         PomodoroGoal findGoal = pomodoroGoalRepository.findById(goalId)
@@ -54,7 +58,7 @@ public class PomodoroService {
         // 사용자 일치 여부 확인
         boolean isUserMatch = findGoal.getUser().getId().equals(id);
         if(!isUserMatch) {
-            throw new IllegalArgumentException("사용자 조회 중 오류가 발생했습니다.");
+            throw new AccessDeniedException("해당 목표를 수정할 권한이 없습니다.");
         }
 
         // 포모도로 목표 수정
@@ -64,8 +68,76 @@ public class PomodoroService {
     }
 
     @Transactional
-    public String delete(Long goalId) {
+    public String delete(Long id, Long goalId) throws AccessDeniedException {
+        PomodoroGoal pomodoroGoal = pomodoroGoalRepository.findById(goalId)
+                .orElseThrow(() -> new NoSuchElementException("해당 목표를 찾을 수 없습니다."));
+
+        if(!pomodoroGoal.getUser().getId().equals(id)) {
+            throw new AccessDeniedException("해당 목표를 삭제할 권한이 없습니다.");
+        }
+
         pomodoroGoalRepository.deleteById(goalId);
         return "목표가 성공적으로 삭제되었습니다.";
+    }
+
+    @Transactional
+    public PomodoroSessionCreateResponse createSession(Long id, @Valid PomodoroSessionCreateRequest dto) {
+        // 포모도로 세션 생성
+        PomodoroSession session = PomodoroSession.builder()
+                .user(userRepository.findById(id).get())
+                .pomodoroGoal(pomodoroGoalRepository.findById(dto.id()).get())
+                .build();
+        pomodoroSessionRepository.save(session);
+
+        return new PomodoroSessionCreateResponse(session.getId(), "포모도로 세션 기록이 시작되었습니다.");
+    }
+
+     // 수정해야함 (dto에 pomodoro_goal_id를 받으면 해당 세션을 종료할 수가 없음.)
+     // 세션은 삭제하는게 아니라 그냥 기록을 유지하는거라서
+    @Transactional
+    public PomodoroSessionEndResponse endSession(@Valid PomodoroSessionEndRequest dto) {
+
+        // durationTime 추가
+        PomodoroSession session = pomodoroSessionRepository.findById(dto.pomodoroSessionId())
+                .orElseThrow(() -> new NoSuchElementException("해당 세션을 찾을 수 없습니다."));
+        session.updateDurationInSeconds(dto.durationInSeconds());
+
+        // 사이클 확인 후 상태 변경
+        if(dto.isCycleCompleted()) {
+            session.updateCycleCompletedStatus(true);
+        }
+
+        return new PomodoroSessionEndResponse(dto.pomodoroSessionId(), "세션 기록이 종료되었습니다.");
+    }
+
+    @Transactional
+    public CoinByPomodoroSessionResponse earnedCoin(Long id, @Valid CoinByPomodoroSessionReqeust dto) {
+
+        // 사용자 조회
+        User user = userRepository.findById(id)
+                .orElseThrow(() -> new NoSuchElementException("사용자를 찾을 수 없습니다."));
+
+        // 마지막 코인 지급일 확인
+        LocalDate today = LocalDate.now();
+        LocalDate lastCoinDate = user.getLastPomodoroCoinDate();
+
+        if(lastCoinDate != null && lastCoinDate.isEqual(today)) {
+            return new CoinByPomodoroSessionResponse(user.getTotalCoins(), "오늘은 이미 코인을 지급받았습니다.");
+        }
+
+        // 포모도로 세션 조회
+        PomodoroSession session = pomodoroSessionRepository.findById(dto.pomodoroSessionId())
+                .orElseThrow(() -> new NoSuchElementException("해당 세션을 찾을 수 없습니다."));
+
+        // isCycleCompleted, isPremiumUser 확인
+        if(session.getIsCycleCompleted() && user.getIsPremium()) {
+            user.addCoins(1);
+            user.updateLastPomodoroCoinDate(today);
+
+            return new CoinByPomodoroSessionResponse(user.getTotalCoins(), "코인이 성공적으로 지급되었습니다.");
+        }
+        else {
+            return new CoinByPomodoroSessionResponse(user.getTotalCoins(), "코인이 지급되지 않았습니다.");
+        }
     }
 }
